refactor note content component to smaller parts
