+++ templates/react-shadcn/src/features/note-widget/components/note-content/hooks/index.ts
@@ -1,0 +1,20 @@
+import { useEffect, useState } from 'react';
+import { useInView } from 'react-intersection-observer';
+
+import { Chunk } from '../types';
+import { parseChunks } from '../utils';
+
+export const useNoteContent = (content: string) => {
+  const [chunks, setChunks] = useState<Chunk[]>([]);
+
+  const { ref, inView } = useInView({
+    threshold: 0,
+    triggerOnce: true,
+  });
+
+  useEffect(() => {
+    setChunks(parseChunks(content));
+  }, [content]);
+
+  return { chunks, ref, inView };
+};
