+++ templates/react-shadcn/src/features/note-widget/components/note-content/types/index.ts
@@ -1,0 +1,6 @@
+export type NostrTypes = 'nprofile' | 'nevent' | 'naddr' | 'nsec' | 'npub' | 'note';
+
+export type Chunk = {
+  type: 'text' | 'image' | 'video' | 'youtube' | 'url' | NostrTypes;
+  content: string;
+};
